pyproject.toml
[project]
name = "code-explainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.75.0",
    "ast-comments>=1.2.3",
    "black>=25.12.0",
    "fastapi>=0.128.0",
    "gitpython>=3.1.46",
    "httpx>=0.28.1",
    "libcst>=1.8.6",
    "mypy>=1.19.1",
    "neo4j>=6.0.3",
    "openai>=2.14.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "ruff>=0.14.10",
    "sentence-transformers>=5.2.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "typer>=0.21.0",
    "uvicorn[standard]>=0.40.0",
]
